refactor!: s/kwctl manifest/kwctl scaffold
Move `manifest` module to `scaffold`.
Move src/manifest.rs to src/scaffold.rs
